requirements.txt
numpy
pandas==1.1.5
pandas-datareader==0.9.0
scikit-learn==0.24.1
sklearn-pandas==2.0.4
torch==1.7.0
transformers==4.2.2